requirements.txt
streamlit
sentence-transformers
faiss-cpu
pdfminer.six
python-docx
nltk
scikit-learn
PyYAML